LayoutTests/fast/hidpi/image-srcset-viewport-modifiers-expected.txt

PASS document.getElementById("foo").clientWidth==100 is true
This test passes if the image below is a blue square, since the srcset is unhandled (uses a viewport modifier that we will skip)
